requirements.txt
Flask==1.1.1
pymongo==3.10.1
pandas==1.0.1
python-dotenv==0.11.0
scipy==1.4.1
nltk==3.4.5
dnspython